Solve day 6 of 2021

+++ 2021/src/d01/mod.rs
@@ -1,20 +1,20 @@
 use itertools::Itertools;
 
-pub fn solve(problem: &[&str]) -> (u32, u32) {
+pub fn solve(problem: &[&str]) -> (u64, u64) {
     let depths = problem
         .iter()
-        .map(|s| s.parse::<u32>().unwrap())
+        .map(|s| s.parse::<u64>().unwrap())
         .collect_vec();
     (solve1(&depths), solve2(&depths))
 }
 
-fn solve1(depths: &[u32]) -> u32 {
+fn solve1(depths: &[u64]) -> u64 {
     let num_increase = depths.windows(2).filter(|w| w[1] > w[0]).count();
-    u32::try_from(num_increase).unwrap()
+    u64::try_from(num_increase).unwrap()
 }
 
-fn solve2(depths: &[u32]) -> u32 {
-    let sums: Vec<u32> = depths.windows(3).map(|w| w.iter().sum()).collect_vec();
+fn solve2(depths: &[u64]) -> u64 {
+    let sums: Vec<u64> = depths.windows(3).map(|w| w.iter().sum()).collect_vec();
     solve1(&sums)
 }
